Harden prediction storage and add joint evidence diagnostics
Version PredictionWindow archives with explicit dense precision, preserve MotionCrafter's native float32 path, add scalable joint covariance calibration diagnostics, and emit deeply validated content-addressed promotion evidence cards alongside the existing held-out failure diagnosis. Add focused regressions, documentation, CI coverage, and source-distribution retention without changing frozen provider-v1/v2 observation contracts.

+++ tests/test_io.py
@@ -39,7 +39,7 @@ def write_problem_bundle(root: Path, problem: SyntheticProblem) -> tuple[Path, P
     manifest_windows = []
     for window in problem.overlap_windows:
         relative = Path("windows") / f"{window.window_id}.npz"
-        window.to_npz(root / relative)
+        window.to_npz(root / relative, storage_dtype="float32")
         manifest_windows.append(
             {
                 "window_id": window.window_id,
